style: move contents module to root
