+++ workflow/service/yes24_service.py
@@ -8,7 +8,7 @@ import requests
 import requests.packages
 from bs4 import BeautifulSoup
 
-from notion_df.object.contents import BlockContents, Heading1BlockContents, Heading2BlockContents, \
+from notion_df.contents import BlockContents, Heading1BlockContents, Heading2BlockContents, \
     Heading3BlockContents, ParagraphBlockContents
 from notion_df.object.rich_text import RichText
